CDHSH-96 Created a backend module for handling requests to the predictions service; fixed the 'config' file import problem (by moving it to utils/).
