+++ services/backend/src/externalServices/index.js
@@ -1,0 +1,5 @@
+const PredictionsService = require("./predictionsService");
+
+module.exports = {
+    PredictionsService
+};
