+++ services/backend/src/hooks/parameterizedHydrate.js
@@ -1,9 +1,9 @@
 const hydrate = require("feathers-sequelize/hooks/hydrate");
 
 const parameterizedHydrate = () => (context) => {
-    const {hydrate} = context.params;
+    const {hydrate: shouldHydrate} = context.params;
 
-    if (hydrate) {
+    if (shouldHydrate) {
         return hydrate().call(context.service, context);
     } else {
         return context;
